+ray class (ray.h) +color function to draw an interpolated background (main.cpp)
