+++ RayTracing/ray.cpp
@@ -1,0 +1,2 @@
+#include "ray.h"
+
